+++ src/message.c
@@ -1295,7 +1295,14 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
     if (token_equals(command, COMMAND_SPACE_FOCUS)) {
         struct selector selector = parse_space_selector(rsp, &message, acting_sid);
         if (selector.did_parse && selector.sid) {
-            space_manager_focus_space(selector.sid);
+            enum space_op_error result = space_manager_focus_space(selector.sid);
+            if (result == SPACE_OP_ERROR_SAME_SPACE) {
+                daemon_fail(rsp, "cannot focus an already focused space.\n");
+            } else if (result == SPACE_OP_ERROR_DISPLAY_IS_ANIMATING) {
+                daemon_fail(rsp, "cannot focus space because the display is in the middle of an animation.\n");
+            } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
+                daemon_fail(rsp, "cannot focus space because mission-control is active.\n");
+            }
         }
     } else if (token_equals(command, COMMAND_SPACE_MOVE)) {
         struct selector selector = parse_space_selector(rsp, &message, acting_sid);
@@ -1305,6 +1312,10 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "cannot move space to itself.\n");
             } else if (result == SPACE_OP_ERROR_SAME_DISPLAY) {
                 daemon_fail(rsp, "cannot move space across display boundaries. use --display instead.\n");
+            } else if (result == SPACE_OP_ERROR_DISPLAY_IS_ANIMATING) {
+                daemon_fail(rsp, "cannot move space because the display is in the middle of an animation.\n");
+            } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
+                daemon_fail(rsp, "cannot move space because mission-control is active.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_SWAP)) {
@@ -1315,6 +1326,10 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "cannot swap space with itself.\n");
             } else if (result == SPACE_OP_ERROR_SAME_DISPLAY) {
                 daemon_fail(rsp, "cannot swap space across display boundaries. use --display instead.\n");
+            } else if (result == SPACE_OP_ERROR_DISPLAY_IS_ANIMATING) {
+                daemon_fail(rsp, "cannot swap space because the display is in the middle of an animation.\n");
+            } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
+                daemon_fail(rsp, "cannot swap space because mission-control is active.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_DISPLAY)) {
@@ -1329,10 +1344,21 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "acting space is the last user-space on the source display and cannot be moved.\n");
             } else if (result == SPACE_OP_ERROR_INVALID_DST) {
                 daemon_fail(rsp, "acting space is already located on the given display.\n");
+            } else if (result == SPACE_OP_ERROR_DISPLAY_IS_ANIMATING) {
+                daemon_fail(rsp, "cannot send space to display because it is in the middle of an animation.\n");
+            } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
+                daemon_fail(rsp, "cannot send space to display because mission-control is active.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_CREATE)) {
-        space_manager_add_space(acting_sid);
+        enum space_op_error result = space_manager_add_space(acting_sid);
+        if (result == SPACE_OP_ERROR_MISSING_SRC) {
+            daemon_fail(rsp, "could not locate the space to act on.\n");
+        } else if (result == SPACE_OP_ERROR_DISPLAY_IS_ANIMATING) {
+            daemon_fail(rsp, "cannot create space because the display is in the middle of an animation.\n");
+        } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
+            daemon_fail(rsp, "cannot create space because mission-control is active.\n");
+        }
     } else if (token_equals(command, COMMAND_SPACE_DESTROY)) {
         enum space_op_error result = space_manager_destroy_space(acting_sid);
         if (result == SPACE_OP_ERROR_MISSING_SRC) {
@@ -1341,6 +1367,10 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
             daemon_fail(rsp, "acting space is the last user-space on the source display and cannot be destroyed.\n");
         } else if (result == SPACE_OP_ERROR_INVALID_TYPE) {
             daemon_fail(rsp, "cannot destroy a macOS fullscreen space.\n");
+        } else if (result == SPACE_OP_ERROR_DISPLAY_IS_ANIMATING) {
+            daemon_fail(rsp, "cannot destroy space because the display is in the middle of an animation.\n");
+        } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
+            daemon_fail(rsp, "cannot destroy space because mission-control is active.\n");
         }
     } else if (token_equals(command, COMMAND_SPACE_BALANCE)) {
         space_manager_balance_space(&g_space_manager, acting_sid);